+++ actix-web-lab/CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ## 0.18.8 - 2022-12-10
 
-- Always add `Content-Encoding` header when using `Sse` as a responder.
+- Always add `Content-Encoding: identity` header when using `Sse` as a responder.
 
 ## 0.18.7 - 2022-12-10